Module Param.Shift

type t

The type of shifting operators.

val id : t

id is the identity (no shifting).

val const : int -> t

const n is the constant shifting fun i -> i + n.

raises Invalid_argument
if n < 0.

val is_id : t -> bool

is_id s checkes whether s is the identity.

val equal : t -> t -> bool

equal x y checks whether x and y are the same operator.

val lt : t -> t -> bool

lt x y checks if x is strictly less than y. Note that trichotomy in general fails for shifting operators.

val le : t -> t -> bool

le x y checks if x is less than or equal to y. Note that trichotomy in general fails for shifting operators.

val compose : t -> t -> t

compose s0 s1 composes the operators s0 and s1, in the texual order.

val dump : Stdlib.Format.formatter -> t -> unit

Ugly printer.